+++ CHANGELOG.rst
@@ -1,12 +1,12 @@
 3.2
 ---
-...
+- Add ``receiver_connections.XXX.csv`` which summarizes connections, disconnections and slow receivers.
 
 3.1
 ---
 - Fix bug resulting in error when log analysis takes more than 10s.
-- Add supprot for .gz files.
-- Add special-case to parse apama-ctrl-* log files that don't end with .log. 
+- Add support for ``.gz`` files.
+- Add special-case to parse ``apama-ctrl-*`` log files that don't end with .log. 
 
 3.0
 ---